Add undo/redo functionality

+++ ui/src/interface/editor/useSelected.tsx
@@ -2,14 +2,14 @@ import { useState } from "react";
 import { Graph } from "../../api/api";
 import { EditorMode } from "./Edit";
 
-export function useSelected(hoveredNodes: number[], setGraph: React.Dispatch<React.SetStateAction<Graph>>, editorMode: EditorMode) {
+export function useSelected(hoveredNodes: number[], graph: Graph, setGraph: (newPresent: Graph, checkpoint?: boolean | undefined) => void, editorMode: EditorMode) {
     const [selectedNodes, setSelectedNodes] = useState<number[]>([]); // these numbers are keys to the nodes map
     const [shiftHeld, setShiftHeld] = useState(false);
 
     function selectionHandleKeyPress(e: React.KeyboardEvent) {
         if (e.key === 'Shift') { // shift must be held for selection of more than one node
             setShiftHeld(e.type === 'keydown')
-        } else if (e.key === 'Delete') {
+        } else if (e.key === 'Delete' && e.type === 'keydown') {
             deleteSelectedNodes();
         }
     }
@@ -44,25 +44,24 @@ export function useSelected(hoveredNodes: number[], setGraph: React.Dispatch<Rea
     }
 
     function deleteSelectedNodes() {
+        console.log("delete selected")
         // Remove selected nodes
-        setGraph((prevGraph) => {
-            // copy old graph
-            const newGraph = {
-                nodes: new Map(prevGraph.nodes),
-                edges: new Map(prevGraph.edges),
-            };
-            // Deselect hovered nodes if any were previously selected and shift is held down
-            selectedNodes.forEach(nodeId => {
-                newGraph.nodes.delete(nodeId) // delete node itself
-                // Delete associated edges
-                for (let [edgeId, edge] of prevGraph.edges) {
-                    if (nodeId === edge[0] || nodeId === edge[1]) {
-                        newGraph.edges.delete(edgeId);
-                    }
+        // copy old graph
+        const newGraph = {
+            nodes: new Map(graph.nodes),
+            edges: new Map(graph.edges),
+        };
+        // Deselect hovered nodes if any were previously selected and shift is held down
+        selectedNodes.forEach(nodeId => {
+            newGraph.nodes.delete(nodeId) // delete node itself
+            // Delete associated edges
+            for (let [edgeId, edge] of graph.edges) {
+                if (nodeId === edge[0] || nodeId === edge[1]) {
+                    newGraph.edges.delete(edgeId);
                 }
-            });
-            return newGraph;
+            }
         });
+        setGraph(newGraph);
     }
 
     return { selectedNodes, deselectAll, selectionHandleKeyPress, selectionHandleClick };
